requirements.txt
django-cors-headers>=3.5.0
django-rest-auth>=0.9.5
pyyaml>=5.3.1